+++ src/views/SubscriptionPage.vue
@@ -2,11 +2,14 @@
   <sidebar-layout>
     <template #content>
       <h1>{{ $t('subscriptions.subscription') }}</h1>
-      <subscription-monero
-        v-if="profile && isMonero()"
-        :profile="profile"
-      ></subscription-monero>
-      <div v-else>{{ $t('subscriptions.no_subscription_info') }}</div>
+      <template v-if="!isLoading">
+        <subscription-monero
+          v-if="profile && isMonero()"
+          :profile="profile"
+        ></subscription-monero>
+        <div v-else>{{ $t('subscriptions.no_subscription_info') }}</div>
+      </template>
+      <loader v-else></loader>
     </template>
   </sidebar-layout>
 </template>
@@ -23,6 +26,7 @@ import {
   Profile,
   ProfilePaymentOption,
 } from "@/api/users"
+import Loader from "@/components/Loader.vue"
 import SidebarLayout from "@/components/SidebarLayout.vue"
 import SubscriptionMonero from "@/components/SubscriptionMonero.vue"
 import { useSubscribe } from "@/composables/subscribe"
@@ -38,6 +42,7 @@ const { setPageTitle } = useTitle()
 
 const profile = ref<Profile | null>(null)
 const subscriptionOption = ref<ProfilePaymentOption | null>(null)
+const isLoading = ref(true)
 
 onMounted(async () => {
   setPageTitle(t("subscriptions.subscription"))
@@ -57,18 +62,21 @@ onMounted(async () => {
     }
   } catch (error: any) {
     if (error.message === "profile not found") {
+      isLoading.value = false
       return
     }
     throw error
   }
   if (isRemoteProfile(_profile) && currentUser.value === null) {
     // Only authenticated users can view remote subscriptions
+    isLoading.value = false
     return
   }
   profile.value = _profile
   // The subscription page is displayed
   // even if current user matches `profile` (as a preview of actual page)
   subscriptionOption.value = getSubscriptionOption(profile.value)
+  isLoading.value = false
 })
 
 function isMonero(): boolean {
@@ -78,3 +86,9 @@ function isMonero(): boolean {
   return isMoneroChain(subscriptionOption.value.chain_id)
 }
 </script>
+
+<style scoped lang="scss">
+.loader {
+  margin: 0 auto;
+}
+</style>