+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Require `Status` entities to have `hidden` attribute.
 - Set `order` parameter to `active` when accessing `/api/v1/directory`.
+- Updated translations ([#130](https://codeberg.org/silverpill/mitra-web/pulls/130)).
 
 ## [4.6.0] - 2025-07-16