+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Improved table styles.
 - Do not enlarge small attached images.
 - Replaced "Subscription expired {date}" with "Subscription expired on {date}".
+- Don't display prompt in post editor when editing existing post.
 
 ## [4.7.0] - 2025-08-01