+++ CHANGELOG.md
@@ -20,6 +20,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't display prompt in post editor when editing existing post.
 - Show empty post editor if markdown source is not available.
 
+### Fixed
+
+- Add space between post header and quote when content is empty.
+
 ## [4.7.0] - 2025-08-01
 
 ### Changed