+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't display prompt in post editor when editing existing post.
 - Show empty post editor if markdown source is not available.
 - Require guest subscriber to enter full Fediverse address at subscription page.
+- Removed link from sender block on subscription page.
 
 ### Fixed