+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Scroll to post when clicking on post timestamp in guest mode.
 - Don't display remote profile pages to unauthenticated users.
 - Open remote site when clicking on "View conversation" button in guest mode.
+- Show loader at subscription page while profile is being retrieved.
 
 ### Fixed